+++ Sources/Room/Room.swift
@@ -81,6 +81,10 @@ class Room {
 
         if let id = group {
             request.group = Int64(id)
+
+            // This is a hack, when a room is created we don't get the group etc back.
+            // However we need it to ensure the admin can't change visibility if the room is in a group.
+            state.group.id = Int64(id)
         }
 
         if let ids = users {
@@ -367,6 +371,10 @@ extension Room: RoomClientDelegate {
             }
 
             self.state.members.append(member)
+        }
+
+        if state.hasGroup {
+            self.state.group = state.group
         }
 
         addMeToState(role: .regular)